requirements.txt
anyio==4.9.0
async-timeout==5.0.1
betterproto==2.0.0b7
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
ffmpy==0.6.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-pubsub==2.31.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mashumaro==3.16
multidict==6.4.4
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
playwright==1.52.0
proto-plus==1.26.1
protobuf==6.31.1
protobuf3-to-dict==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-socks==2.7.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
tenacity==9.1.2
TikTokApi==7.1.0
TikTokLive==6.5.3
typing_extensions==4.14.0
urllib3==2.4.0
websockets==15.0.1
websockets_proxy==0.1.3
Werkzeug==3.1.3
zipp==3.23.0
